requirements.txt
aws-cdk.core==1.94.1
aws-cdk.assets==1.94.1
aws-cdk.aws-iam==1.94.1
aws-cdk.aws-ec2==1.94.1
aws-cdk.aws-lambda-python==1.94.1
aws-cdk.aws-apigateway==1.94.1
-e .
